[#15] Address code comment: refactor & fix auth response
